refactored files and folders
