+++ client/src/App.js
@@ -3,8 +3,10 @@ import React from 'react';
 import { BrowserRouter as Router, Route} from "react-router-dom";
 
 import NavBar from './component/NavBar/NavBar';
-import Signup from './component/Signup/Signup';
-import Login from './component/Login/Login'
+import Signup from './component/Users/Signup/Signup';
+import Login from './component/Users/Login/Login'
+import Home from './component/Home';
+import Dashboard from './component/Dashboard';
 
 import "./App.css";
 
@@ -15,9 +17,10 @@ const App = () => {
     <>
       <NavBar />
     <Router>
-      <Route path="/" component={Home} />
-      <Route path="/signup" component={Signup} />
-      <Route path="/login" component={Login} />
+      <Route path="/" component={Home} exact />
+      <Route path="/signup" component={Signup} exact />
+      <Route path="/login" component={Login} exact />
+      <Route path="/dashboard" component={Dashboard}/>
     </Router>
     </>
   );
