+++ client/src/component/Dashboard/index.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Dashboard() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
