+++ client/src/component/Home/index.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Home() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
