+++ client/src/component/NewsFeed/TechNews/index.js
